add-check-service-cost-on-backend

+++ src/main/java/ru/borisova/boostingservice/models/Service.java
@@ -15,6 +15,6 @@ public class Service {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     public Long id;
     public String name;
-    public Integer cost;
-    public Integer discount;
+    public Float cost;
+    public Float discount;
 }
